titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/complexity-code/data-access-complexity-code/src/lib/+state/complexity-code.selectors.ts
import { createSelector } from '@ngrx/store';
import { selectComplexityCodeState } from './complexity-code.reducer';

const selectcomplexityCodeList = createSelector(
  selectComplexityCodeState,
  state => state.compexityCodeList
);

const selectTotalElements = createSelector(
  selectComplexityCodeState,
  state => state.totalElements
);
const selectIsloading = createSelector(
  selectComplexityCodeState,
  state => state.isLoading
);

const selectError = createSelector(
  selectComplexityCodeState,
  state => state.error
);

const selectComplexityCode = createSelector(
  selectComplexityCodeState,
  state => state.complexityCode
);
const selectHasSaved = createSelector(
  selectComplexityCodeState,
  state => state.hasSaved
);
const selectHasUpdated = createSelector(
  selectComplexityCodeState,
  state => state.hasUpdated
);

export const complexityCodeSelector = {
  selectComplexityCode,
  selectcomplexityCodeList,
  selectTotalElements,
  selectIsloading,
  selectError,
  selectHasSaved,
  selectHasUpdated
};
